requirements.txt
# Multi-Agent Debate System Dependencies
langchain>=0.3.0
langgraph>=0.2.0
langchain-openai>=0.2.0
pydantic>=2.0
python-dotenv>=1.0.0
streamlit>=1.40.0
